# static/swagger.yaml
openapi: 3.0.0
info:
  description: 	╰( ͡° ͜ʖ ͡° )つ──☆*:・ﾟ
  version: 1.0.0
  title: FHIR API
servers:
- url: "/"
tags:
- name: Patient Request
  description: Example API for requesting and returning patient requests
- name: Schedule Request
  description: Example API for requesting and returning patient requests
- name: Electronic Medical Record Request
  description: Example API for requesting and returning patient's EMR
paths:
  "/create_patient":
    post:
      tags:
      - Patient Request
      summary: Create a new patient
      requestBody:
        description: Patient Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/patientRequestPostBody"
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/patient/{id}":
    get:
      tags:
        - Patient Request
      summary: Returns patient with given ID
      parameters:
        - in: path
          name: id
          required: true
          description: Patient Request id
          schema:
            "$ref": "#/components/schemas/id"
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/components/schemas/patientRequest"
  "/get_patient":
    post:
      tags:
      - Patient Request
      summary: Get patient
      requestBody:
        description: Patient Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/patientGetRequest"
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/create_schedule":
    post:
      tags:
      - Schedule Request
      summary: Create a new schedule
      requestBody:
        description: Schedule Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/scheduleRequestPostBody"
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/create_slot":
    post:
      tags:
      - Schedule Request
      summary: Create a new slot
      requestBody:
        description: Slot Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/slotRequestPostBody"
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/create_appointment":
    post:
      tags:
      - Schedule Request
      summary: Create a new appointment
      requestBody:
        description: Appointment Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/appointmentRequestPostBody"
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/get_appointments":
    post:
      tags:
      - Schedule Request
      summary: Get appointments of patient
      requestBody:
        description: Appointments Request
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/appointmentsRequest"
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/get_slots":
    post:
      tags:
      - Schedule Request
      summary: Get slots of practitioner
      requestBody:
        description: Slots Request
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/slotsRequest"
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/create_observations":
    post:
      tags:
      - Electronic Medical Record Request
      summary: Create new observations for a patient
      requestBody:
        description: EMR Request Post Object
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                  $ref: '#/components/schemas/observationBody'
      produces:
      - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/create_diagnoses":
    post:
      tags:
        - Electronic Medical Record Request
      summary: Create new diagnoses for a patient
      requestBody:
        description: EMR Request Post Object
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/diagnosticBody'
      produces:
        - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/create_medications":
    post:
      tags:
        - Electronic Medical Record Request
      summary: Create new medications for a patient
      requestBody:
        description: EMR Request Post Object
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/medicationBody'
      produces:
        - application/json
      responses:
        '201':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/get_observations":
    post:
      tags:
        - Electronic Medical Record Request
      summary: Get existing observations for a patient
      requestBody:
        description: EMR Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/patientGetRequest"
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/get_diagnoses":
    post:
      tags:
        - Electronic Medical Record Request
      summary: Get existing diagnoses for a patient
      requestBody:
        description: EMR Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/patientGetRequest"
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
  "/get_medications":
    post:
      tags:
        - Electronic Medical Record Request
      summary: Get existing medications for a patient
      requestBody:
        description: EMR Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/patientGetRequest"
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        '400':
          description: Failed. Bad post data.
components:
  schemas:
    id:
      properties:
        uuid:
          type: string
    patientRequestPostBody:
      type: object
      properties:
        policyNumber:
          type: integer
          format: integer
        name:
          type: string
          format: string
        family:
          type: string
          format: string
        gender:
          type: string
          format: string
        birthDate:
          type: string
          format: YYYY-MM-DD
        phoneNumber:
          type: string
          format: "+xxxxxxxxxxx"
        email:
          type: string
          format: x@x.x
        address:
          type: string
          format: string
        city:
          type: string
          format: string
        state:
          type: string
          format: string
    patientGetRequest:
      type: object
      properties:
        policyNumber:
          type: integer
          format: integer
    scheduleRequestPostBody:
      type: object
      properties:
        serviceCategoryCode:
          type: integer
          format: integer
        serviceCategory:
          type: string
          format: string
        serviceTypeCode:
          type: integer
          format: integer
        serviceType:
          type: string
          format: string
        specialtyCode:
          type: integer
          format: integer
        specialty:
          type: string
          format: string
        Practitioner:
          type: string
          format: string
        PractitionerID:
          type: string
          format: string
        Location:
          type: string
          format: string
        LocationID:
          type: string
          format: string
        StartDate:
          type: string
          format: YYYY-MM-DDTHH:MM:SSZ
        EndDate:
          type: string
          format: YYYY-MM-DDTHH:MM:SSZ
    slotRequestPostBody:
      type: object
      properties:
        serviceCategoryCode:
          type: integer
          format: integer
        serviceCategory:
          type: string
          format: string
        serviceTypeCode:
          type: integer
          format: integer
        serviceType:
          type: string
          format: string
        specialtyCode:
          type: integer
          format: integer
        specialty:
          type: string
          format: string
        appointmentTypeCode:
          type: string
          format: string
        appointmentType:
          type: string
          format: string
        ScheduleID:
          type: string
          format: string
        StartDate:
          type: string
          format: YYYY-MM-DDTHH:MM:SSZ
        EndDate:
          type: string
          format: YYYY-MM-DDTHH:MM:SSZ
    appointmentRequestPostBody:
      type: object
      properties:
        slotID:
          type: string
          format: string
        policyNumber:
          type: string
          format: string
    appointmentsRequest:
      type: object
      properties:
        policyNumber:
          type: string
          format: string
    slotsRequest:
      type: object
      properties:
        practitionerID:
          type: string
          format: string
        startDate:
          type: string
          format: YYYY-MM-DDTHH:MM:SSZ
        endDate:
          type: string
          format: YYYY-MM-DDTHH:MM:SSZ
    observationBody:
      type: object
      properties:
        ObservationStatus:
          type: string
          format: string
        ObservationText:
          type: string
          format: string
        policyNumber:
          type: string
          format: string
        EffectiveDateTime:
          type: string
          format:
            oneOf:
              - YYYY
              - YYYY-MM
              - YYYY-MM-DD
              - YYYY-MM-DDThh:mm:ss+zz:zz
        PatientName:
          type: string
          format: string
        QuantityValue:
          type: number
          format: duble
        QuantityUnit:
          type: string
          format: string
        Interpretation:
          type: string
          format: string
    medicationBody:
      type: object
      properties:
        ObservationID:
          type: string
          format: string
        MedicationStatus:
          type: string
          format: string
        MedicationConcept:
          type: string
          format: string
        DateAsserted:
          type: string
          format:
            oneOf:
              - YYYY
              - YYYY-MM
              - YYYY-MM-DD
              - YYYY-MM-DDThh:mm:ss+zz:zz
        PractitionerID:
          type: string
          format: string
        policyNumber:
          type: string
          format: string
        Dosage:
          type: string
          format: string
        AdditionalInstruction:
          type: array
          items:
            type: string
    diagnosticBody:
      type: object
      properties:
        VisitReason:
          type: string
          format: string
        ObservationID:
          type: string
          format: string
        PractitionerID:
          type: string
          format: string
        policyNumber:
          type: string
          format: string
        EffectiveDateTime:
          type: string
          format:
            oneOf:
              - YYYY
              - YYYY-MM
              - YYYY-MM-DD
              - YYYY-MM-DDThh:mm:ss+zz:zz
        Conclusion:
          type: string
          format: string
      diagnosticBody:
        type: object
        properties:
          VisitReason:
            type: string
            format: string
          ObservationID:
            type: string
            format: string
          PractitionerID:
            type: string
            format: string
          policyNumber:
            type: string
            format: string
          EffectiveDateTime:
            type: string
            format:
              oneOf:
                - YYYY
                - YYYY-MM
                - YYYY-MM-DD
                - YYYY-MM-DDThh:mm:ss+zz:zz
          Conclusion:
            type: string
            format: string